api/requirements.txt
bottle==0.12.13
psycopg2-binary==2.7.5
